Specify the minimum Perl version explicitly
I noticed that `Perl::MinimumVersion` showed that no minimum version was
set explicitly and that the minimum Perl version required for this dist
was version 5.6.  This change specifies the version requirement
explicitly in the module itself as well as in the `Build.PL` script so
that the version requirement can be checked at build time.  This change
also corrects the `meta_yml_declares_perl_version` CPANTS issue.