Fix GitHub action builds on Windows
I noticed that the GitHub workflow for Windows was failing and decided
to work out why.  The build logs showed that `Module::Pluggable` was
failing its tests, which seemed rather odd.  After much wailing and
knashing of teeth (not really), I worked out that the
`shogo82148/actions-setup-perl@v1` was causing the problem.  Exactly *why*
it was causing this problem is unclear; nevertheless, omitting this
action for the Windows-based builds enabled all of the dependencies to
be installed via `cpanm --installdeps .`

Since the `shogo82148/actions-setup-perl@v1` action was causing the
problem, it seemd best to split the jobs into two: one for the
Unix-based builds and one for the Windows-based builds.  I've also added
more debugging output to the `cpanm` calls so that if some dependency
*does* fail in the future, then the detailed build information will also
be present as part of the GitHub action's build log.