Extend README to contain installation, usage and dev docs
When viewing the README on GitHub, I noticed that there wasn't that much
information in it, hence I thought it'd be good for future users and
contributors to have more information about the module up front and
available directly on GitHub.  I've tried to cover module installation
and usage as well as how to contribute to the module as best I can.  At
the very least, this is a beginning :-)